open_env_project/customer_support_env/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "support_ticket_env"
version = "1.0.0"
description = "Customer Support Ticket Resolution Environment - OpenEnv Compliant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.7.0",
    "gradio>=4.20.0",
    "openai>=1.0.0",
    "numpy>=1.26.0",
    "pytest>=8.0.0",
]

[project.scripts]
support-server = "customer_support_env.app:main"
